feat: PaymentService to realize payments and rollback failure payments.

+++ src/main/java/br/com/fpnbr/paymentservice/domain/Payment.java
@@ -35,7 +35,7 @@ public class Payment {
     private LocalDateTime updatedAt;
 
     @Column(name = "total_items", nullable = false)
-    private Integer totalItems;
+    private Long totalItems;
 
     @Column(name = "total_amount", nullable = false)
     private BigDecimal totalAmount;
